% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_age_birth_inci.R
\name{check_age_birth_inci}
\alias{check_age_birth_inci}
\title{Check consistency between age and dates.}
\usage{
check_age_birth_inci(age, birthda = NULL, inciden = NULL, quiet = FALSE)
}
\arguments{
\item{age}{An integer vector indicating the ages of the cancer cases.}

\item{birthda}{A Date vector indicating the birth dates of the cancer cases.}

\item{inciden}{A Date vector indicating the incidence dates of the
cancer cases.}

\item{quiet}{Logical. If TRUE, suppresses output messages.}
}
\value{
A list object of class \code{check} with the following elements:
\itemize{
\item \code{check_item}: A character string indicating the item or combination of
items being checked.
\item \code{check_standard}: A character string specifying the standard used.
\item \code{check_version}: A character string specifying the version of the standard.
\item \code{data}: A list containing the input data.
\item \code{check}: A logical vector indicating whether each code passed the checks.
\item \code{type}: A numeric vector categorizing validation results: \code{1} for correct,
\code{2} for Warning, \code{3} for Error.
\item \code{error_code}: A character vector containing error or warning codes.
\item \code{error_desc}: A character vector containing descriptive messages for each
error or warning.
}
}
\description{
Check consistency between age and dates.
}
\examples{
age <- c(58, 59, 60)
birth <- c("1960-08-24", "1956-09-24", "1971-06-30")
onset <- c("2024-09-10", "2024-08-10", "2023-09-11")
res <- check_age_birth_inci(age, birth, onset)
print(res)
}
