% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_age_site_morp.R
\name{check_age_site_morp}
\alias{check_age_site_morp}
\title{Check consistency between data items of age, site, and morp}
\usage{
check_age_site_morp(age, topo, morp)
}
\arguments{
\item{age}{Age}

\item{topo}{Cancer site}

\item{morp}{Morphology}
}
\value{
Class of 'check'
}
\description{
Check consistency between data items of age, site, and morp
}
\examples{
age <- c(30)
topo <- c("C61.9")
morp <- c(8140)
}
