% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_age_topo_morp.R
\name{check_age_topo_morp}
\alias{check_age_topo_morp}
\title{Check consistency between data items of age, site, and morp}
\usage{
check_age_topo_morp(age, topo, morp, std = "iacr", quiet = FALSE)
}
\arguments{
\item{age}{An integer vector indicating the ages of the cancer cases.}

\item{topo}{A character vector containing the topography part of ICD-O-3
codes (e.g., "C50.9").}

\item{morp}{A character vector containing the morphology part of ICD-O-3
codes (e.g., "8000" or \code{c("8000", "8070", "8140")}).}

\item{std}{A character string indicating the standard to use.
Options are "iarc" (default) or "encr".}

\item{quiet}{Logical. If TRUE, suppresses output messages.}
}
\value{
A list object of class \code{check} with the following elements:
\itemize{
\item \code{check_item}: A character string indicating the item or combination of
items being checked.
\item \code{check_standard}: A character string specifying the standard used.
\item \code{check_version}: A character string specifying the version of the standard.
\item \code{data}: A list containing the input data.
\item \code{check}: A logical vector indicating whether each code passed the checks.
\item \code{type}: A numeric vector categorizing validation results: \code{1} for correct,
\code{2} for Warning, \code{3} for Error.
\item \code{error_code}: A character vector containing error or warning codes.
\item \code{error_desc}: A character vector containing descriptive messages for each
error or warning.
}
}
\description{
Check consistency between data items of age, site, and morp
}
\examples{
age <- c(30)
topo <- c("C61.9")
morp <- c(8140)
res <- check_age_topo_morp(age, topo, morp)
print(res)
}
