% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_basi.R
\name{check_basi}
\alias{check_basi}
\title{Check the Basis of diagnosis code}
\usage{
check_basi(basi, quiet = FALSE)
}
\arguments{
\item{basi}{A character vector indicating the basis of diagnosis for the
cancer cases.}

\item{quiet}{Logical. If TRUE, suppresses output messages.}
}
\value{
A list object of class \code{check} with the following elements:
\itemize{
\item \code{check_item}: A character string indicating the item or combination of
items being checked.
\item \code{check_standard}: A character string specifying the standard used.
\item \code{check_version}: A character string specifying the version of the standard.
\item \code{data}: A list containing the input data.
\item \code{check}: A logical vector indicating whether each code passed the checks.
\item \code{type}: A numeric vector categorizing validation results: \code{1} for correct,
\code{2} for Warning, \code{3} for Error.
\item \code{error_code}: A character vector containing error or warning codes.
\item \code{error_desc}: A character vector containing descriptive messages for each
error or warning.
}
}
\description{
Check the Basis of diagnosis code
}
\examples{
basi <- c(1, 2, 4, 5, 6, 7, 9, 3)
res <- check_basi(basi)
print(res)
}
