% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_basi_morp.R
\name{check_basi_morp}
\alias{check_basi_morp}
\title{Check consistency between basis of diagnosis and morphology code}
\usage{
check_basi_morp(
  basi,
  morp,
  topo = NULL,
  beha = NULL,
  std = "iacr",
  quiet = FALSE
)
}
\arguments{
\item{basi}{A character vector indicating the basis of diagnosis for the
cancer cases.}

\item{morp}{A character vector containing the morphology part of ICD-O-3
codes (e.g., "8000" or \code{c("8000", "8070", "8140")}).}

\item{topo}{A character vector containing the topography part of ICD-O-3
codes (e.g., "C50.9").}

\item{beha}{A character vector containing the behavior codes of the ICD-O-3
codes, (e.g., "3" or \code{c("1", "2", "5")}).}

\item{std}{A character string indicating the standard to use.
Options are "iarc" (default) or "encr".}

\item{quiet}{Logical. If TRUE, suppresses output messages.}
}
\value{
A list object of class \code{check} with the following elements:
\itemize{
\item \code{check_item}: A character string indicating the item or combination of
items being checked.
\item \code{check_standard}: A character string specifying the standard used.
\item \code{check_version}: A character string specifying the version of the standard.
\item \code{data}: A list containing the input data.
\item \code{check}: A logical vector indicating whether each code passed the checks.
\item \code{type}: A numeric vector categorizing validation results: \code{1} for correct,
\code{2} for Warning, \code{3} for Error.
\item \code{error_code}: A character vector containing error or warning codes.
\item \code{error_desc}: A character vector containing descriptive messages for each
error or warning.
}
}
\description{
Check consistency between basis of diagnosis and morphology code
}
\examples{
 basi <- c(1, 1, 1, 1, 5, 7, 2)
 beha <- c(3, 3, 3, 3, 3, 3, 3)
 morp <- c(8000, 8720, 9120, 9510, 8023, 8001, 9180)
 topo <- c("C15.2","C15.2", "C44.0", "C69.1", "C42.1", "C15.2", "C15.2")
 res <- check_basi_morp(basi, morp, topo, beha)
 print(res)
}
