% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_beha.R
\name{check_beha}
\alias{check_beha}
\title{Verify if the Behavior codes for ICDO3 comply with the ICDO3 coding rules.}
\usage{
check_beha(beha, quiet = FALSE)
}
\arguments{
\item{beha}{A character vector containing the behavior codes of the ICD-O-3
codes, (e.g., "3" or \code{c("1", "2", "5")}).}

\item{quiet}{Logical. If TRUE, suppresses output messages.}
}
\value{
A list object of class \code{check} with the following elements:
\itemize{
\item \code{check_item}: A character string indicating the item or combination of
items being checked.
\item \code{check_standard}: A character string specifying the standard used.
\item \code{check_version}: A character string specifying the version of the standard.
\item \code{data}: A list containing the input data.
\item \code{check}: A logical vector indicating whether each code passed the checks.
\item \code{type}: A numeric vector categorizing validation results: \code{1} for correct,
\code{2} for Warning, \code{3} for Error.
\item \code{error_code}: A character vector containing error or warning codes.
\item \code{error_desc}: A character vector containing descriptive messages for each
error or warning.
}
}
\description{
Verify if the Behavior codes for ICDO3 comply with the ICDO3 coding rules.
}
\examples{
behas <- c(0, 1, 2, 3, 9, 6, 8)
res <- check_beha(behas)
print(res)
}
