% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_beha_site.R
\name{check_beha_site}
\alias{check_beha_site}
\title{Check consistency between behaviour and site code}
\usage{
check_beha_site(beha, site)
}
\arguments{
\item{beha}{Behaviour code}

\item{site}{Cancer site code}
}
\value{
class of 'check'
}
\description{
Check consistency between behaviour and site code
}
\examples{
beha <- c(2, 2, 2, 2)
site <- c("C40.9", "C49.1", "C70.0", "C72.9")
check_beha_site(beha, site)
}
