% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_beha_topo.R
\name{check_beha_topo}
\alias{check_beha_topo}
\title{Check consistency between behavior and site code}
\usage{
check_beha_topo(beha, topo, quiet = FALSE)
}
\arguments{
\item{beha}{A character vector containing the behavior codes of the ICD-O-3
codes, (e.g., "3" or \code{c("1", "2", "5")}).}

\item{topo}{A character vector containing the topography part of ICD-O-3
codes (e.g., "C50.9").}

\item{quiet}{Logical. If TRUE, suppresses output messages.}
}
\value{
A list object of class \code{check} with the following elements:
\itemize{
\item \code{check_item}: A character string indicating the item or combination of
items being checked.
\item \code{check_standard}: A character string specifying the standard used.
\item \code{check_version}: A character string specifying the version of the standard.
\item \code{data}: A list containing the input data.
\item \code{check}: A logical vector indicating whether each code passed the checks.
\item \code{type}: A numeric vector categorizing validation results: \code{1} for correct,
\code{2} for Warning, \code{3} for Error.
\item \code{error_code}: A character vector containing error or warning codes.
\item \code{error_desc}: A character vector containing descriptive messages for each
error or warning.
}
}
\description{
Check consistency between behavior and site code
}
\examples{
beha <- c(2, 2, 2, 2)
site <- c("C40.9", "C49.1", "C70.0", "C72.9")
res <- check_beha_topo(beha, site)
print(res)
}
