% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_birth_inci_last.R
\name{check_birth_inci_last}
\alias{check_birth_inci_last}
\title{Check consistency between birth date incidence and Date of last known vital status}
\usage{
check_birth_inci_last(birthda, inciden, lastcontact, quiet = FALSE)
}
\arguments{
\item{birthda}{A Date vector indicating the birth dates of the cancer cases.}

\item{inciden}{A Date vector indicating the incidence dates of the
cancer cases.}

\item{lastcontact}{A Date vector indicating the date of last follow-up,
which is the most recent date on which the vital status of a cancer
patient is known. If the patient has died, this corresponds to the
date of death.}

\item{quiet}{Logical. If TRUE, suppresses output messages.}
}
\value{
A list object of class \code{check} with the following elements:
\itemize{
\item \code{check_item}: A character string indicating the item or combination of
items being checked.
\item \code{check_standard}: A character string specifying the standard used.
\item \code{check_version}: A character string specifying the version of the standard.
\item \code{data}: A list containing the input data.
\item \code{check}: A logical vector indicating whether each code passed the checks.
\item \code{type}: A numeric vector categorizing validation results: \code{1} for correct,
\code{2} for Warning, \code{3} for Error.
\item \code{error_code}: A character vector containing error or warning codes.
\item \code{error_desc}: A character vector containing descriptive messages for each
error or warning.
}
}
\description{
Check consistency between birth date incidence and Date of last known vital status
}
