% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_grade_morp.R
\name{check_grade_morp}
\alias{check_grade_morp}
\title{Check consistency between grade and morphology}
\usage{
check_grade_morp(grade, morp)
}
\arguments{
\item{grade}{grade code}

\item{morp}{morphology code}
}
\value{
class of 'check'
}
\description{
Check consistency between grade and morphology
}
\examples{
grade <- c(3, 5)
morp <- c(9705, 9714)
}
