% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_morp.R
\name{check_morp}
\alias{check_morp}
\title{Verify if the Morphology codes for ICDO3 comply with the ICDO3
coding rules.}
\usage{
check_morp(morp, quiet = FALSE)
}
\arguments{
\item{morp}{A character vector containing the morphology part of ICD-O-3
codes (e.g., "8000" or \code{c("8000", "8070", "8140")}).}

\item{quiet}{Logical. If TRUE, suppresses output messages.}
}
\value{
A list object of class \code{check} with the following elements:
\itemize{
\item \code{check_item}: A character string indicating the item or combination of
items being checked.
\item \code{check_standard}: A character string specifying the standard used.
\item \code{check_version}: A character string specifying the version of the standard.
\item \code{data}: A list containing the input data.
\item \code{check}: A logical vector indicating whether each code passed the checks.
\item \code{type}: A numeric vector categorizing validation results: \code{1} for correct,
\code{2} for Warning, \code{3} for Error.
\item \code{error_code}: A character vector containing error or warning codes.
\item \code{error_desc}: A character vector containing descriptive messages for each
error or warning.
}
}
\description{
Verify if the Morphology codes for ICDO3 comply with the ICDO3
coding rules.
}
\examples{
morps <- c(9053, 8000, 8170, 8050, 8111)
res <- check_morp(morps)
print(res)

}
