% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sex.R
\name{check_sex}
\alias{check_sex}
\title{Check validation of sex code}
\usage{
check_sex(sex)
}
\arguments{
\item{sex}{sex code}
}
\value{
class of 'check'
}
\description{
Check validation of sex code
}
\examples{
sex <- c(1, 2, 3, 1, 2)
check_sex(sex)
}
