% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sex_morp.R
\name{check_sex_morp}
\alias{check_sex_morp}
\title{Check consistency between sex and morphology code}
\usage{
check_sex_morp(sex = c(1, 2), morp = c(8905, 8080))
}
\arguments{
\item{sex}{sex code}

\item{morp}{morphology code}
}
\value{
class of 'check'
}
\description{
Check consistency between sex and morphology code
}
\examples{
sex <- c(1, 2, 1, 1)
morp <- c(8905, 8080, 9000, 9300)
}
