% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sex_topo.R
\name{check_sex_topo}
\alias{check_sex_topo}
\title{Check consistency between sex and site code}
\usage{
check_sex_topo(sex, topo, std = "iacr", quiet = FALSE)
}
\arguments{
\item{sex}{A character vector indicating the sex of the cancer cases,
where \code{"1"} represents male and \code{"2"} represents female.}

\item{topo}{A character vector containing the topography part of ICD-O-3
codes (e.g., "C50.9").}

\item{std}{A character string indicating the standard to use.
Options are "iarc" (default) or "encr".}

\item{quiet}{Logical. If TRUE, suppresses output messages.}
}
\value{
A list object of class \code{check} with the following elements:
\itemize{
\item \code{check_item}: A character string indicating the item or combination of
items being checked.
\item \code{check_standard}: A character string specifying the standard used.
\item \code{check_version}: A character string specifying the version of the standard.
\item \code{data}: A list containing the input data.
\item \code{check}: A logical vector indicating whether each code passed the checks.
\item \code{type}: A numeric vector categorizing validation results: \code{1} for correct,
\code{2} for Warning, \code{3} for Error.
\item \code{error_code}: A character vector containing error or warning codes.
\item \code{error_desc}: A character vector containing descriptive messages for each
error or warning.
}
}
\description{
Check consistency between sex and site code
}
\examples{
sex <- c(1, 1, 1, 2, 2, 2)
topo <- c("C51.0", "C53.0", "C57.4", "C60.0", "C60.9", "C63.9")
res <- check_sex_topo(sex, topo)
print(res)
}
