% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_topo.R
\name{check_topo}
\alias{check_topo}
\title{Verify if the topographic site codes for ICDO3 comply with the ICDO3
coding rules.}
\usage{
check_topo(x)
}
\arguments{
\item{x}{Character vector contains the topographic site codes for ICDO3.}
}
\value{
check_topo returns an object of class of 'check'.
}
\description{
Verify if the topographic site codes for ICDO3 comply with the ICDO3
coding rules.
}
\examples{
topos <- c("C50.9", "16.2", "C151", "33.2")
check_topo(topos)
}
