% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_topo_morp_beha_grad.R
\name{check_topo_morp_beha_grad}
\alias{check_topo_morp_beha_grad}
\title{Check consistency between Topography, Morphology, Behavior, and Grade.}
\usage{
check_topo_morp_beha_grad(topo, morp, beha, grad, quiet = FALSE)
}
\arguments{
\item{topo}{A character vector containing the topography part of ICD-O-3
codes (e.g., "C50.9").}

\item{morp}{A character vector containing the morphology part of ICD-O-3
codes (e.g., "8000" or \code{c("8000", "8070", "8140")}).}

\item{beha}{A character vector containing the behavior codes of the ICD-O-3
codes, (e.g., "3" or \code{c("1", "2", "5")}).}

\item{grad}{A character vector containing the grade codes of the ICD-O-3
codes, (e.g., \code{c("1", "2", "3", "4", "9")}).}

\item{quiet}{Logical. If TRUE, suppresses output messages.}
}
\value{
A list object of class \code{check} with the following elements:
\itemize{
\item \code{check_item}: A character string indicating the item or combination of
items being checked.
\item \code{check_standard}: A character string specifying the standard used.
\item \code{check_version}: A character string specifying the version of the standard.
\item \code{data}: A list containing the input data.
\item \code{check}: A logical vector indicating whether each code passed the checks.
\item \code{type}: A numeric vector categorizing validation results: \code{1} for correct,
\code{2} for Warning, \code{3} for Error.
\item \code{error_code}: A character vector containing error or warning codes.
\item \code{error_desc}: A character vector containing descriptive messages for each
error or warning.
}
}
\description{
Check consistency between Topography, Morphology, Behavior, and Grade.
}
