% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_trib.R
\name{check_trib}
\alias{check_trib}
\title{Verify codes of nationalities whether comply with GB3304-91 in China}
\usage{
check_trib(trib, quiet = FALSE)
}
\arguments{
\item{trib}{A character vector containing the trib codes used in
Chinese cancer registry.}

\item{quiet}{Logical. If TRUE, suppresses output messages.}
}
\value{
A list object of class \code{check} with the following elements:
\itemize{
\item \code{check_item}: A character string indicating the item or combination of
items being checked.
\item \code{check_standard}: A character string specifying the standard used.
\item \code{check_version}: A character string specifying the version of the standard.
\item \code{data}: A list containing the input data.
\item \code{check}: A logical vector indicating whether each code passed the checks.
\item \code{type}: A numeric vector categorizing validation results: \code{1} for correct,
\code{2} for Warning, \code{3} for Error.
\item \code{error_code}: A character vector containing error or warning codes.
\item \code{error_desc}: A character vector containing descriptive messages for each
error or warning.
}
}
\description{
Verify codes of nationalities whether comply with GB3304-91 in China
}
