% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{comm_return_value}
\alias{comm_return_value}
\title{Common Return Value for check Functions}
\value{
A list object of class \code{check} containing:
\itemize{
\item \code{check_item}: Character contains item or items combination
being checked.
\item \code{check_standard}: Character contains the standard used
\item \code{check_version}:  Character contains the version of the
standard used.
\item \code{data}: A list contains the input data.
\item \code{check}: A logical vector indicating whether each code passed
the checks.
\item \code{type}: A numeric vector categorizing validation results,
\itemize{
\item \code{1}: Correct
\item \code{2}: Warning
\item \code{3}: Error
}
\item \code{error_code}: A character vector providing error or warning codes.
\item \code{error_desc}: A numeric vector contains description codes for
specific information.
}
}
\description{
Common Return Value for check Functions
}
\keyword{internal}
