% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{compress_morp}
\alias{compress_morp}
\title{Compress a morphology code vector into a compact range string}
\usage{
compress_morp(x, sep = ",", pattern = "\%04d", conn = "-", min_range = 3)
}
\arguments{
\item{x}{A numeric or integer vector of morphology codes.}

\item{sep}{A string used to separate values in the output. Default is ",".}

\item{pattern}{A format string passed to \code{sprintf()} (e.g., "\%04d").}

\item{conn}{A character used to connect range start and end. Default is "-".}

\item{min_range}{Integer. Minimum length of a run to compress. Default is 3.}
}
\value{
A character string with compressed ranges.
}
\description{
Converts a numeric vector (e.g., ICD-O-3 morphology codes) into a compact
character string using hyphens for consecutive values and commas for
separation.
}
\examples{
compress_morp(c(8000, 8001, 8002, 8003, 9594, 8006))
compress_morp(c(8000, 8001, 8002), min_range = 4)
}
