% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{compress_topo}
\alias{compress_topo}
\title{Compress ICD-O-3 topography codes}
\usage{
compress_topo(topo, sep = ",", conn = "-", min_range = 3)
}
\arguments{
\item{topo}{A character vector of ICD-O-3 topography codes (e.g., "C17.0").}

\item{sep}{A string used to separate values in the output. Default is (",").}

\item{conn}{A string used to connect compressed ranges. Default is "-".}

\item{min_range}{Integer. Minimum number of sequential codes to compress into a range.}
}
\value{
A character string of compressed topography codes.
}
\description{
If all subcodes from CXX.0 to CXX.9 are present, compress to "CXX".
Otherwise, list each code individually. Additionally, compress sequential
top-level codes (e.g., "C01,C02,C03" → "C01-C03") when possible.
}
\examples{
compress_topo(c("C17.0", "C17.1", "C17.2", "C17.3", "C17.4", 
                "C17.5", "C17.6", "C17.7", "C17.8", "C17.9"))

compress_topo(c("C17.0", "C17.2", "C17.9"))

compress_topo(c("C01.0", "C02.1", "C03.9", "C04.2", "C04.9"))

compress_topo(gen_topo(10:19))
}
