% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crcheck.R
\name{cr_check}
\alias{cr_check}
\alias{cr_check.data.frame}
\alias{cr_check.canreg}
\title{Conduct quality check for cancer registration data.}
\usage{
cr_check(
  x,
  std = "iacr",
  registr = "registr",
  sex = "sex",
  age = "age",
  birthda = "birthda",
  inciden = "inciden",
  topo = "topo",
  morp = "morp",
  beha = "beha",
  grad = "grad",
  icd10 = "icd10",
  ...
)

\method{cr_check}{data.frame}(
  x,
  std = "iacr",
  registr = "registr",
  sex = "sex",
  age = "age",
  birthda = "birthda",
  inciden = "inciden",
  topo = "topo",
  morp = "morp",
  beha = "beha",
  grad = "grad",
  icd10 = "icd10",
  ...
)

\method{cr_check}{canreg}(x, ...)
}
\arguments{
\item{x}{Object contains cancer registration data}

\item{std}{Character, indicate the standard used to check the validation of
the item or items combination, options are "iacr" or "encr".}

\item{registr}{The record number of cancer case}

\item{sex}{The birth sex of cancer case}

\item{age}{The age of cancer case}

\item{birthda}{The birth date of cancer case}

\item{inciden}{The incidence date of cancer case}

\item{topo}{The cancer site code of cancer case coded as ICDO3.2}

\item{morp}{The Morphology code of cancer case, coded as ICDO3.2}

\item{beha}{The behaviour code of cancer case, coded as ICDO3.2}

\item{grad}{The grade code of cancer code, coded as ICDO3.2}

\item{icd10}{The ICD10 code corresponding to ICD3.2}

\item{...}{Other parameters}
}
\value{
A data frame contain the check result.

A data frame contain the check result.

A data frame with check results.
}
\description{
Conduct quality check for cancer registration data.

Conduct quality check for cancer registration data.

Conduct quality check for cancer registration data.
}
