% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crcheck-package.R
\docType{package}
\name{crcheck-package}
\alias{crcheck}
\alias{crcheck-package}
\title{crcheck: Checking the Validity and Internal Consistency}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Perform internal consistency checks on individual variables or combinations of variables collected in cancer registries, in accordance with the rules of the International Agency for Research on Cancer (IARC) / the International Association of Cancer Registries (IACR).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gigu003/crcheck}
}

}
\author{
\strong{Maintainer}: Qiong Chen \email{chenq08@126.com} (\href{https://orcid.org/0000-0003-2401-0046}{ORCID}) (Department of Cancer Epidemiology, The Affiliated Cancer Hospital of Zhengzhou University & Henan Cancer Hospital, Henan Cancer Registry in China) [copyright holder]

}
\keyword{internal}
