% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{gen_icd10}
\alias{gen_icd10}
\title{Generate valid ICD10 codes from ranges of numbers}
\usage{
gen_icd10(x, init = "C", na_drop = TRUE)
}
\arguments{
\item{x}{Numbers}

\item{init}{Initial letter, "C" or "D"}

\item{na_drop}{Logical value indicates whether drop NA values or not.}
}
\value{
Vector of ICD10 codes.
}
\description{
Generate valid ICD10 codes from ranges of numbers
}
\examples{
gen_icd10(180:229)
gen_icd10(220:229)
}
