% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{gen_morp}
\alias{gen_morp}
\title{Generate valid ICDO3 morphology codes from range of numbers.}
\usage{
gen_morp(x, na_drop = TRUE)
}
\arguments{
\item{x}{Numbers}

\item{na_drop}{Logical value indicates whether drop NA values or not.}
}
\value{
Vector of valid ICDO3 morphology codes.
}
\description{
Generate valid ICDO3 morphology codes from range of numbers.
}
\examples{
gen_morp(8000:8090)
gen_morp(8500:8700)
}
