% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_topo.R
\name{gen_topo}
\alias{gen_topo}
\title{generate legal ICDO3 Topo code}
\usage{
gen_topo(x, na.drop = TRUE)
}
\arguments{
\item{x}{Number between 0 and 809}

\item{na.drop}{Drop illegal code or not.}
}
\value{
A vector of legeal TOPO codes.
}
\description{
generate legal ICDO3 Topo code
}
\examples{
 gen_topo(1:169)
}
