% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICDO3toICD10.R
\name{ICDO3toICD10}
\alias{ICDO3toICD10}
\title{Convert ICDO3 code to ICD10 code.}
\usage{
ICDO3toICD10(topo, morp, beha)
}
\arguments{
\item{topo}{Topography code in ICDO3.}

\item{morp}{Morphology code in ICDO3.}

\item{beha}{Behavior code in ICDO3.}
}
\value{
Vector of ICD10 codes.
}
\description{
Convert ICDO3 code to ICD10 code.
}
