% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{param_basi}
\alias{param_basi}
\title{Basis}
\arguments{
\item{basi}{A character vector to be checked whether comply with the basis
coding rules of cancer diagnosis, which is a one-digit numerical code
of (0, 1, 2, 4, 5, 7, 9). Invalid codes
will be flagged during validation.}
}
\description{
Basis
}
\keyword{internal}
