% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{param_beha}
\alias{param_beha}
\title{Behaviour}
\arguments{
\item{beha}{A character vector to be checked whether comply with the ICD-O-3
behaviour coding rules, which is a one-digit numerical code (0, 1, 2,
3, 6, 9), and indicate the cancer behaviour. Invalid codes will be
flagged during validation. For more details about the rules, refer
to the official ICD-O-3 coding guidelines:
\url{http://www.iacr.com.fr/index.php?Itemid=577}.}
}
\description{
Behaviour
}
\keyword{internal}
