% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{param_grad}
\alias{param_grad}
\title{Grade}
\arguments{
\item{grad}{A character vector to be checked whether comply with the ICD-O-3
grade coding rules, which is a one-digit numerical code ranging from
1 to 9, and indicate the grade of cancer development. Invalid codes
will be flagged during validation. For more details about the rules,
refer to the official ICD-O-3 coding guidelines:
\url{http://www.iacr.com.fr/index.php?Itemid=577}.}
}
\description{
Grade
}
\keyword{internal}
