% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{param_icd10}
\alias{param_icd10}
\title{ICD10}
\arguments{
\item{icd10}{A character vector to be checked whether comply with the ICD10
coding rules in tumor part. Invalid codes will be flagged during
validation. For more details about the rules, refer to the official
ICD10 coding guidelines: \url{https://icd.who.int/browse10/2019/en}.}
}
\description{
ICD10
}
\keyword{internal}
