% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{param_morp}
\alias{param_morp}
\title{Morphology}
\arguments{
\item{morp}{A character vector to be checked whether comply with the ICD-O-3
morphology coding rules, which is a 4-digit numerical code ranging
from 8000 to 9993 indicate the cancer histology, but it is not
continuous, some 4-digit codes within this range are not valid codes,
e.g.(8977, 8978), and you can use \code{gen_morp(8000:9993)} to list all
valid morphology codes. Invalid codes will be flagged during
validation. For more details about the rules, refer to the official
ICD-O-3 coding guidelines: \url{http://www.iacr.com.fr/index.php?Itemid=577}.}
}
\description{
Morphology
}
\keyword{internal}
