% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{param_occu}
\alias{param_occu}
\title{Occupation}
\arguments{
\item{occu}{Character vector to be checked whether comply with code of
occupation of GB/T2261.4 in China.}
}
\description{
Occupation
}
\keyword{internal}
