% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{param_show_summary}
\alias{param_show_summary}
\title{show_summary}
\arguments{
\item{show_summary}{Logical, whether to display a summary of the validation
results. Defaults to \code{TRUE}. If \code{FALSE}, only the result
object will be returned.}
}
\description{
show_summary
}
\keyword{internal}
