% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{param_std}
\alias{param_std}
\title{Standard}
\arguments{
\item{std}{Character, indicate the standard used to check the validation of
the item or items combination, options are "iacr" or "encr".}
}
\description{
Standard
}
\keyword{internal}
