% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{param_topo}
\alias{param_topo}
\title{Topography}
\arguments{
\item{topo}{A character vector to be checked whether comply with the ICD-O-3
topography coding rules, which indicate the topography cancer sites,
e.g.(C50.9). Invalid codes (e.g., missing decimal points or
non-conforming formats) will be flagged during validation. For more
details about the rules, refer to the official ICD-O-3 coding
guidelines: \url{http://www.iacr.com.fr/index.php?Itemid=577}.}
}
\description{
Topography
}
\keyword{internal}
