% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{param_trib}
\alias{param_trib}
\title{Nationalities}
\arguments{
\item{trib}{Character vector to be checked whether comply with code of
nationalities of GB3304-91 in China.}
}
\description{
Nationalities
}
\keyword{internal}
