% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.check}
\alias{print.check}
\title{Print list of check result}
\usage{
\method{print}{check}(x, ...)
}
\arguments{
\item{x}{Object of class 'check'}

\item{...}{Other parameters.}
}
\value{
A data frame.
}
\description{
Print list of check result
}
