% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_morp}
\alias{recode_morp}
\title{Recode the morphology code in a string}
\usage{
recode_morp(x)
}
\arguments{
\item{x}{A string contains morphology code}
}
\value{
Vector
}
\description{
Recode the morphology code in a string
}
