% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_topo}
\alias{recode_topo}
\title{Recode the topography code in a string}
\usage{
recode_topo(x)
}
\arguments{
\item{x}{A string contains topography code}
}
\value{
Vector
}
\description{
Recode the topography code in a string
}
