% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_morp_beha.R
\name{show_morp_beha}
\alias{show_morp_beha}
\title{Show morphology–behavior rules}
\usage{
show_morp_beha(morp, beha = c(1:3, 9))
}
\arguments{
\item{morp}{A character vector containing the morphology part of ICD-O-3
codes (e.g., "8000" or \code{c("8000", "8070", "8140")}).}

\item{beha}{A character vector containing the behavior codes of the ICD-O-3
codes, (e.g., "3" or \code{c("1", "2", "5")}).}
}
\value{
A character vector of valid morphology–behavior combinations
(e.g., "8000/3").
}
\description{
Show valid combinations of ICD-O morphology and behavior codes according to
ICD-O-3 rules.
}
\examples{
show_morp_beha(morp = 8000:8140)
show_morp_beha(morp = c(8000, 8010), beha = c(2, 3))
}
