% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.check}
\alias{summary.check}
\title{Show the summary of the check result}
\usage{
\method{summary}{check}(object, ...)
}
\arguments{
\item{object}{check class}

\item{...}{Other parameters}
}
\value{
summary
}
\description{
Show the summary of the check result
}
