% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_icd10.R
\name{check_icd10}
\alias{check_icd10}
\title{Check the ICD10 code in the tumor part.}
\usage{
check_icd10(icd10)
}
\arguments{
\item{icd10}{The ICD10 codes.}
}
\value{
check_icd10 returns an object of class of 'check'.
}
\description{
Check the ICD10 code in the tumor part.
}
\examples{

icd10 <- c("C15.2", "C17.0", "C19.0", "C34.9", "D15.2")
res <- check_icd10(icd10)
}
