% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_id.R
\name{check_id}
\alias{check_id}
\title{Check if the Chinese ID number complies with the GB 11643-1999 standard.}
\usage{
check_id(x, return = "logical")
}
\arguments{
\item{x}{A character vector containing Chinese ID numbers to be checked for
compliance with the GB 11643-1999 standard.}

\item{return}{A character string used to specific the type of result
returned. Options are 'logical', 'date', 'areacode', 'formatted'.
Options include:
\itemize{
\item{'logical': Logical vector (default). Indicates whether
each input ID number complies with the standard (TRUE for
compliant, FALSE for non-compliant).}
\item{'date': Date vector representing birth date contained in
the ID number.}
\item{'areacode': Character vector containing administrative
region codes extracted from the ID number.}
\item{'formatted': Character vector containing formatted ID
numbers.}
}}
}
\value{
A logical vector indicating whether each input ID number complies
with the standard (TRUE for compliant, FALSE for non-compliant),
or other types of information based on the 'return' parameter.
}
\description{
Check if the Chinese ID number complies with the GB 11643-1999 standard.
}
\examples{
ids <- c(
  "412726198407240038", "110101199003070975",
  "310101199001010101", "440101199102030303"
)
check_id(ids)
check_id(ids, return = "date")
check_id(ids, return = "areacode")
check_id(ids, return = "formatted")

}
