% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_morp_beha.R
\name{check_morp_beha}
\alias{check_morp_beha}
\title{Check the morphology and behavior code in ICDO3}
\usage{
check_morp_beha(morp, beha)
}
\arguments{
\item{morp}{Morphology codes.}

\item{beha}{Behavior codes.}
}
\value{
check_morp returns an object of class of 'check'.
}
\description{
Check the morphology and behavior code in ICDO3
}
\examples{
morp <- c(8000, 8010, 8140, 8533, 9000)
beha <- c(3, 3, 3, 3, 1)
}
