% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sex_site.R
\name{check_sex_topo}
\alias{check_sex_topo}
\title{Check consistency between sex and site code}
\usage{
check_sex_topo(sex, topo)
}
\arguments{
\item{sex}{sex code}

\item{topo}{cancer site code}
}
\value{
class of 'check'
}
\description{
Check consistency between sex and site code
}
\examples{
sex <- c(1, 1, 1, 2, 2, 2)
topo <- c("C51.0", "C53.0", "C57.4", "C60.0", "C60.9", "C63.9")
}
