% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_topo_morp.R
\name{check_topo_morp}
\alias{check_topo_morp}
\title{Check the combination of TOPO and MORP.}
\usage{
check_topo_morp(topo = c("C15.2", "C33.9"), morp = c(8240, 8012))
}
\arguments{
\item{topo}{TOPO code of ICDO3}

\item{morp}{MOPR code of ICDO3}
}
\value{
Class of check and topo/morp
}
\description{
Check the combination of TOPO and MORP.
}
\examples{
topo <- c("C15.1", "C06.9", "C16.2", "C34.9")
morp <- c(8140, 8500, 8500, 8600)
check_topo_morp(topo, morp)
}
