% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{gen_topo}
\alias{gen_topo}
\title{Generate valid ICDO3 topography codes.}
\usage{
gen_topo(x, na_drop = TRUE)
}
\arguments{
\item{x}{Number between 0 and 809}

\item{na_drop}{Logical value indicates whether drop NA values or not.}
}
\value{
A vector of valid ICDO3 topography codes.
}
\description{
Generate valid ICDO3 topography codes.
}
\examples{
gen_topo(160:169)
gen_topo(170:179)
gen_topo(330:349)
}
